requirements.txt
telebot
requests
BeautifulSoup4
pillow
selenium
webdriver-manager
